github: Enable clang based Windows and Linux build
This commit adds a build job on GitHub.

Any time something is pushed to the "clang-build" branch it will trigger
a build and store (and host) the artifacts.

Normally on such branches when working alone I just force-push. I don't
know the correct workflow if a lot of people works together on the same
project.